move input detection hook so it's generally accessible
