+++ studio/src/hooks/useInputFocusDetection.ts
