+++ studio/src/pages/RequestDetailsPage/hooks/index.ts
@@ -3,4 +3,3 @@ export { usePagination } from "./usePagination";
 export { useReplayRequest } from "./useReplayRequest";
 export { useMostRecentRequest } from "./useMostRecentRequest";
 export { useShouldReplay } from "./useShouldReplay";
-export { useInputFocusDetection } from "./useInputFocusDetection";
